requirements.txt
langgraph
langchain
langchain-openai
pydantic
python-dotenv
faiss-cpu
numpy
